pyproject.toml
[tool.poetry]
name = "bloodhound-core"
version = "0.1.0"
description = "Apache Bloodhound Issue Tracker core functionality"
authors = ["Apache Bloodhound <dev@bloodhound.apache.org>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.2.2"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.20.0"
drf-nested-routers = "^0.93.3"

[tool.poetry.dev-dependencies]
selenium = "^3.141.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
